move schemacompare out of dacfx folders
